@remotion/studio: Move outline flag back to timeline selection
