+++ packages/studio/src/components/Timeline/TimelineSelection.tsx
@@ -10,11 +10,11 @@ import React, {
 	type CSSProperties,
 } from 'react';
 import {StudioServerConnectionCtx} from '../../helpers/client-id';
+import {BACKGROUND} from '../../helpers/colors';
 import type {
 	SequenceNodePathInfo,
 	TrackWithHash,
 } from '../../helpers/get-timeline-sequence-sort-key';
-import {ENABLE_OUTLINES} from '../../helpers/outline-flags';
 import {TIMELINE_PADDING} from '../../helpers/timeline-layout';
 import {timelineNodePathInfoToKey} from '../../helpers/timeline-node-path-key';
 import {useKeybinding} from '../../helpers/use-keybinding';
@@ -63,6 +63,11 @@ export const getTimelineSelectedTrackHighlightStyle = (
 
 export const SELECTION_ENABLED = false;
 export const TIMELINE_TOP_DRAG = false;
+export const ENABLE_OUTLINES = false;
+export const TIMELINE_BACKGROUND = ENABLE_OUTLINES ? '#0F1113' : '#111';
+export const TIMELINE_TICKS_BACKGROUND = ENABLE_OUTLINES
+	? BACKGROUND
+	: TIMELINE_BACKGROUND;
 export const EASING_SELECTION_ENABLED = false;
 
 type TimelineSelectionBase = {
