--- packages/studio/src/helpers/outline-flags.ts
@@ -1,1 +1,0 @@
-export const ENABLE_OUTLINES = false;
